additions for linux-x86_64 hosted AROS
git-svn-id: https://svn.aros.org/svn/aros/trunk@23158 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ tools/FlexCat/flexcat.c
@@ -476,7 +476,7 @@ unsigned long SwapLong1234(unsigned long r)
 int SwapChoose(void)
 {
 unsigned short w;
-unsigned long  d;
+unsigned int d;
 
   strncpy((char *)&w, "\1\2", 2);
   strncpy((char *)&d, "\1\2\3\4", 4);
